Feature audio cd identify (#64)
* Enhance audio disc logging

Enhanced the logging methodology for audio discs. If an audio disc is
identified the musicbrainz or cddb disc ID is used as a logging file
name. This makes multipe rip processes easier to identify and
troubleshoot, and puts audio ripping logs in position to be cleaned by
the existing log-cleaning routine.

* upgrade audio disc logging capabilities

Enhance the logging methodology for audio discs. If an audio disc is
identified the musicbrainz or cddb disc ID is used as a logging file
name. This makes multipe rip processes easier to identify and
troubleshoot, and puts audio ripping logs in position to be cleaned by
the existing log-cleaning routine.