Added logic for trying to parse raw DVD title.
Modified omdb api logic in case there is no year to start.

Added 16X9 to title things to strip.

Handled a rare condition where a DVD does not actually have a name/label.

Stop OMDB lookups when we cannot get a proper title from the DVD Label

Added a logging error when a mainfeature is not found.