Updates forms.py
With the data-validator on, this the ChangeParamForm was rendered useless. Setting the tickbox to be manditory defeats the purpose  of listing the feature. PLUS with no error reporting on the form, the user thinks the form is broken.

Also added the custom title form. Custom Title form is intended for use for DVD's that have multiple disk (al ah Lord of the rings), Or TV Shows where you want to rip 1 disk after another or inparalle.

This way you don't have to manually go back and rename the folder after it's done.