Update handbrake.py
Added a check loop.
Only start handbrake when there are less handbrakes running than user defined.
This way you can que up jobs to run.
With CPU encoding, there is a point of deminishing returns on concurrent encodes, and on GPU encoding, some video cards are locked to the number of encodes possible at once (It will crash out your job 1/4 of the way in).
This was also a way to stopping dupliate jobs from crashing out viable jobs.