Enhance MakeMKV Calls with Parser and Error Handling
With this update, we've improved our MakeMKV integration by introducing
a better parser that handles errors more robustly and provides greater
clarity. The module is now also Pylint compatible, making it easier to
maintain. To reduce noise in the logs, we've removed links to stdout
logfiles, instead directing error reports through Python's built-in
logger. This change reduces the file size of logs for users on info log
level or higher. Additionally, when an internal error occurs that is
likely to be expected due to a known failure, it will be marked as
critical.

These Changes will allow further improvements on working with the
MakeMKV output. We introduce modern python concepts like dataclass and
enums.