Use Enum for CDS (IOctl Drive Status)
Use an Enum to aggregate the possible outputs of the Kernel drive status
into readable output. Use tooltips to show this to the user. Make the
drive status a callable function instead of a (magic) property to
indicate that this function call might take some time and is not pure
state conversion (which is what properties should do).