Preserve Job Association on Database deletion
If a job in the database is deleted, we cannot know for sure if the
association between job.devpath and job_id_previous is actually the
right choice here. It may be that the server restarted in-between
deletions and the mounts have changed. It may also be that the current
job is still running during deletions. In general, this call adds
unnecessary obfuscation to where the job_id_previous is actually changed.

Adding a better (Exception) message to the job info page if a job
is requested but missing and hiding unassociated job ids completley from
the drive settings page.