[BUGFIX] debian 12 install script (#1109)
* Created new Branch for new feature: "Create a Bash Installation Script to install on Bare Metal in a Debian 12 Installation."

Issue #1070

Create Script and function "Has_Sudo_Privileges()"

* Added Descriptive Error message and exit command to shell script if the user does not have Sudo privileges.

* Correct an error where I failed to understand that BASH does not preserve Global Variable values outside of functions.

Added function to create arm Group.

* Logic for creating User and group completed.

Need to seperate it into more readable functions.

* Refactor code for User and group deletion.
Comment Code for legibility.

* [Automated] Increment Version

* Refactored the code, added more comments.

Added Building and Installing of MakeMKV.

Fixed a bug where the script would not complete if run as sudo.

* [Automated] Increment Version

* Create function to install ARM Deb Package Dependencies.

Create function to test for the availability of the contrib package, if that test fails, exit the script.

* [Automated] Increment Version

* Added a function to clone the repository at the latest Arm Release or to checkout the repository, if existing, at the latest ARM Release.

* [Automated] Increment Version

* Correct Code Editor Warnings

* [Automated] Increment Version

* Added User Disclaimer, Informing the user that
1. This installation method is unsupported.
2. MakeMKV is in Beta and free to use while in Beta, and provide a link to MakeMKV purchase page.
3. Provide a link to the ARM MIT licence

* [Automated] Increment Version

* Return Version Number to what it should be.

GitHub Worklow was updating the version number automatically.  Returned the Version Number to what it should be after disabling the GitHub Workflow.

* Create Function to create the Python Virtual Environment for ARM and install the Python Requirements.

* To not break previous installs, create the following files.
A new Udev Rules file.
A new Amr.service file.
A new arm_venv_wrapper.sh file

Updated Debian12Installer.sh file to use the newly created files. Also added a testing flag to pull from testing repo and not production repo.

* Change the location of the created SystemD Service from /etc/systemd/system to /lib/systemd/system

* Added an early check to confirm the sudo package is installed.  (By default sudo is not installed in the Proxmox LXC Debian 12 container image...)

Added a stub function for running the post-install setup UI

* Comment out SUDO_FLAG and all attached code, this was a unnecessary duplication of code that could lead to hard to find errors and bugs.

Added verification for the installation of the sudo package.

Added verification for the installation of the lsb-release package and the ability for the script to install it if missing.

Succesfully tested in a Proxmox LXC container enviroment.

* Deleted the commented-out code that enabled the abadonned optional SUDO script call.  (The ability to call the script with or without SUDO...  Now the script needs to be called with SUDO or as root)

* Add numerous TODO messages to keep track of changes and modifications yet to be made...

* Remove and IF statement dealing with the deprecated SudoFlag variable.

* Completed the Launch Setup Function.

Added script options instead of the hardcoded Test Flag of earlier versions.

* Fixed an issue where we create a new user but fail to add a password due to user error (fail to enter match the password 3 times). The user and group arm would be created but no password would be set on the account. Now if the user fails to enter a password, the newly crated account is deleted.

if the arm user account is pre-existing, give the user the opportunity to change it's password.

* Added some checks for the existence of the repo and / or tag options passed to the script.  Now the script exits with and error if the fork or tag was not found.

* To keep it simpler.  Changed the script from trying to update the github repo of a prior installation to deleting any existing /opt/arm directory and clonning a fresh repository from the chosen fork and tag.  Give the user the option to decline to proceed.  Copy any existing config files to backups.

* add lsb-release package to the initial apt install call therefore keeping the apt update call to just one.

* Tested script with Debian 11 and Debian 12.

More testing is required to attempt to test more variables.

* Update TODOs

* Complete rewrite to reflect the new installation script for Debian and OMV.

This wiki entry is still a draft.

* Fix a glaring omission regarding the Officially supported installation method.

* Fix a typo

* Remove Completed TODOs

* Added test for Amrui.service, if active, exit install script.
If innactive, warn the user that proceeding is not recommended but allow for the user to continue the process.

Moved test for presence of /opt/amr directory ealier, before any installation is completed, to allow the user to abandon installation earlier and only make minimal changes to the system.

Added link to MakeMKV-Info wiki page.

* Fixed a typo in ErrorCode 214.

* Update Setting-up-ARM-script-(Debian-OMV).md

- Added Pre-Installation Steps
- Added 2 new error codes
- Modified the formatting of the Note at the top of the page regarding the deprecated nature of this installation method.

* Create Virtualization-Proxmox.md

A brief pros and cons evaluation of which Proxmox method to use.

* Fixed the wiki updates, after the long pause in development of this feature.

The original Debian Install file was renamed for only OMV, and a new file, was created for Debian.  Restored the OMV file to be identical to main, and moved my changes to the Debian file.

* Add a TODO to fix the problem presented when installing ARM using Proxmox LXC containers.

* Allow ARM to bind protected ports (ei: 1~1024)  This only applies to the Debian Install Script and not the Docker setup.

* Fix armUI Service Found when it did not exist error.

Update the Wiki Entry page, add a note about the length (in time) of the installation, remove mentioned support for OMV.

Update User Disclaimer in script, to reflect current date and to remove OMV support mention.

Update Version Number

* Fix Wiki entry regarding the valid part number, now arm can be installed using any port number even the lower levels from 1 to 1024.  Remove Error code 210 about port number being in the restricted range.

* Update cautionary disclaimer at the top of the wiki page.

Add Intended Use section to the wiki page.

* Remove a file that was never intended to be part of this feature update.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: wolfy <wolfy@shitwolfymakes.com>
Co-authored-by: Mtech <62650032+microtechno9000@users.noreply.github.com>