BugFix: sqlalchemy backref
sqlalchemy backrefs needs only be defined in one place. The way it was
done, created 3 independent backrefs which was a bit confusing and not
working properly when updating one of them. We do not need a
backref for the last job association since the job does not know if it
is the current or previous job for the drive. Only the job needs to
know, what drive it was assigned to. This simplifies the backrefs to just
one (active drive <--> active job) instead of three.