Only delete mdisc association at program start
With the current implementation, mdisc could get deleted by the UI
settings "scan for drives" operation. If this is done in-between a rip
that requires the makemkvcon disc association, the job will fail due to
missing mdisc id. Only deleting the mdisc once on UI startup ensures
that it will get filled by the first call to makemkvcon info and not get
emptied afterwards.