Add more known Errors from MakeMKV
Add the error code for a failure in disc opening. This probably means
that the disc is beeing read by another process. We differentiate
between fatal errors which raise (MakeMKV exit status is unreliably
sometimes) and errors which are only directed to a higher log level to
get user awareness.