Organize Possible Job Status and changes.
Created an enum for JobState to keep better track of the different
possible job states throughout the project. It also prevents new jobs to
get created when any ripping (audio/video) is active. A job is not
released when its status changes. The associated drive keeps track if
the drive can be safely re-used by another job.

Issue #1411