Handle Non-Existing PIDs while releasing job
This wraps the psutil functions into its own try-catch loop. After the
psutil loop, a process is either terminated or an error occurred that is
not yet handled.
If the process is terminated, we return "true". If an error occurred, we
revert the database so that the job will still be visible and the user
can try to kill it once more.

The prerequisite for releasing the drive from the current job is that the
associated process is terminated. Otherwise the drive is not released.
Releasing the drive from the job is completely handled by job.eject().
No call to a (possibly non-existing) associated drive is done on the
abandon-layer.