[BUGFIX] prep_mkv not working (#1509)
* BugFix: prep_mkv not working

The prep_mkv function is internally calling the update_key.sh script.
Due to shell=True, the command is never executing. This also fixes the
command output to stdout, adds proper Error handling in the bash script
and increases the verbosity level of the script on errors so that users
on info level know what is happening. It also defines Known Errors both
in the bash script and in makemkv.py.

* Version Increment